// src/Web/MeAjudaAi.Web.Provider/lib/api/generated/types.gen.ts
export type MeAjudaAiModulesProvidersApplicationDtosAddressDto = {
    street?: string | null;
    number?: string | null;
    complement?: string | null;
    neighborhood?: string | null;
    city?: string | null;
    state?: string | null;
    zipCode?: string | null;
    country?: string | null;
};

/**
 * DTO para perfil empresarial.
 */
export type MeAjudaAiModulesProvidersApplicationDtosBusinessProfileDto = {
    legalName?: string | null;
    fantasyName?: string | null;
    description?: string | null;
    contactInfo?: MeAjudaAiModulesProvidersApplicationDtosContactInfoDto;
    primaryAddress?: MeAjudaAiModulesProvidersApplicationDtosAddressDto;
    showAddressToClient?: boolean;
};

/**
 * DTO para informações de contato.
 */
export type MeAjudaAiModulesProvidersApplicationDtosContactInfoDto = {
    email?: string | null;
    phoneNumber?: string | null;
    website?: string | null;
    additionalPhones?: Array<string> | null;
};
// ...
export type MeAjudaAiModulesProvidersApplicationDtosProviderServiceDto = {
    serviceId?: string;
    serviceName?: string | null;
};
// ...
export type MeAjudaAiModulesProvidersApplicationDtosRequestsUpdateProviderProfileRequest = {
    /**
     * Nome do prestador.
     */
    name?: string | null;
    businessProfile?: MeAjudaAiModulesProvidersApplicationDtosBusinessProfileDto;
    /**
     * Lista de serviços oferecidos pelo prestador. Pode ser nulo se não houver alteração.
     */
    services?: Array<MeAjudaAiModulesProvidersApplicationDtosProviderServiceDto> | null;
};